FlightTermination/requirements.txt
future==1.0.0
lxml==5.2.1
pymavlink==2.4.41
pyserial==3.5
python-dotenv==1.0.1
